refactor(login): PR 리뷰 반영, Google 로그인 파일 구조 개선
