+++ src/assets/icons/GoogleIcon.tsx
@@ -1,0 +1,15 @@
+export const GoogleIcon = ({ className }: { className?: string }) => (
+  <svg
+    className={className}
+    viewBox='0 0 24 24'
+    aria-hidden='true'
+  >
+    <path
+      d='M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26
+        1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74
+        3.28-8.09z'
+      fill='#4285F4'
+    />
+    {/* ... 나머지 path들 ... */}
+  </svg>
+);
